+++ src/js/render-myLibrary.js
@@ -1,21 +1,14 @@
 //? імпорт функцій 
-import { getFromStorage } from "./localStorage";
-import { getMovieByID } from "./getMovieByID";
-import layOutListOfMyLib from "./layOutListMyLibrary";
 import { Notify } from 'notiflix';
 import { clearPage } from './fetch_by_keyword';
-let arrMoviesQueue = [];
-let arrMoviesWatch = [];
-export {clickBtnWatched, clickBtnQueue}
-//! функція рендеру розмітки бібліотеки
+import { getFromStorageAndRender } from "./getFromStorageAndRender";
 
-    // buttonQueue.addEventListener('click', );
-  
+export {clickBtnWatched, clickBtnQueue}
+
     // ! функція кліку по кнопці ватч
  function clickBtnWatched(e) {
       e.preventDefault();
       clearPage();
-      arrMoviesWatch = [];
       // додавання і забирання класів
       document.querySelector('#watched').classList.add('btn-nav-active');
       document.querySelector('#queue').classList.remove('btn-nav-active');
@@ -25,17 +18,7 @@ export {clickBtnWatched, clickBtnQueue}
         Notify.failure('Sorry... You did not add any movie to your Watchlist');
         return;
       }
-      getFromStorage('idWatched').map(id => {
-        getMovieByID(id)
-          .then(movie => {
-            arrMoviesWatch.push(movie);
-            return arrMoviesWatch;
-          })
-          .then(result => {
-            clearPage()
-            layOutListOfMyLib(result)
-          });
-      });
+      getFromStorageAndRender();
     }
   
     // ! функція кліку по кнопці черга
@@ -51,15 +34,6 @@ export {clickBtnWatched, clickBtnQueue}
         return;
       }
       // доставання ід масива та їх перебор з локального сховища
-      getFromStorage('idQueue').map(id => {
-        getMovieByID(id)
-          .then(movie => {
-            arrMoviesQueue = [];
-            arrMoviesQueue.push(movie);
-            return arrMoviesQueue;
-          })
-          .then(layOutListOfMyLib);
-      });
+      getFromStorageAndRender();
     }
 
-
